+++ lib/endpoints/app/jobs.ex
@@ -1,8 +1,7 @@
 defmodule SHEx.Endpoints.App.Jobs do
   import SHEx.Endpoints.Guards
 
-  alias SHEx.Endpoints.App
-  alias SHEx.HttpAdapters.Default, as: DefaultAdapter
+  alias SHEx.Endpoints.{App, Helpers}
   alias SHEx.HttpAdapter.RequestConfig
 
   @base_url "https://app.scrapinghub.com/api"
@@ -12,7 +11,7 @@ defmodule SHEx.Endpoints.App.Jobs do
 
   def run(api_key, project_id, spider_name, params \\ [], opts \\ [])
       when is_api_key(api_key)
-      when is_project_id(project_id)
+      when is_id(project_id)
       when is_binary(spider_name)
       when is_list(params)
       when is_list(opts) do
@@ -31,7 +30,7 @@ defmodule SHEx.Endpoints.App.Jobs do
       |> Map.put(:body, body)
       |> Map.put(:opts, opts)
       |> Map.put(:url, "#{@base_url}/run.json")
-      |> make_request()
+      |> Helpers.make_request()
     else
       error -> {:error, error}
     end
@@ -39,10 +38,10 @@ defmodule SHEx.Endpoints.App.Jobs do
 
   def list(api_key, project_id, params \\ [], opts \\ [])
       when is_api_key(api_key)
-      when is_project_id(project_id)
+      when is_id(project_id)
       when is_list(params)
       when is_list(opts) do
-    with :ok <- validate_params(params, @valid_list_params ++ App.pagination_params()),
+    with :ok <- Helpers.validate_params(params, @valid_list_params ++ App.pagination_params()),
          :ok <- params |> Keyword.get(:state) |> validate_state() do
       query = [{:project, project_id} | params] |> URI.encode_query()
 
@@ -50,7 +49,7 @@ defmodule SHEx.Endpoints.App.Jobs do
       |> Map.put(:api_key, api_key)
       |> Map.put(:opts, opts)
       |> Map.put(:url, "#{@base_url}/jobs/list.json?#{query}")
-      |> make_request()
+      |> Helpers.make_request()
     else
       {:invalid_params, _} = error -> {:error, error}
     end
@@ -58,17 +57,17 @@ defmodule SHEx.Endpoints.App.Jobs do
 
   def update(api_key, project_id, job_or_jobs, params \\ [], opts \\ [])
       when is_api_key(api_key)
-      when is_project_id(project_id)
-      when is_job_id(job_or_jobs) or is_list(job_or_jobs)
+      when is_id(project_id)
+      when is_id(job_or_jobs) or is_list(job_or_jobs)
       when is_list(params)
       when is_list(opts) do
-    with :ok <- validate_params(params, @valid_update_params) do
+    with :ok <- Helpers.validate_params(params, @valid_update_params) do
       request = prepare_basic_request(api_key, project_id, job_or_jobs, opts)
 
       request
       |> Map.put(:url, "#{@base_url}/jobs/update.json")
       |> Map.put(:body, request.body ++ params)
-      |> make_request()
+      |> Helpers.make_request()
     else
       {:invalid_params, _} = error -> {:error, error}
     end
@@ -76,22 +75,22 @@ defmodule SHEx.Endpoints.App.Jobs do
 
   def delete(api_key, project_id, job_or_jobs, opts \\ [])
       when is_api_key(api_key)
-      when is_project_id(project_id)
-      when is_job_id(job_or_jobs) or is_list(job_or_jobs)
+      when is_id(project_id)
+      when is_id(job_or_jobs) or is_list(job_or_jobs)
       when is_list(opts) do
     prepare_basic_request(api_key, project_id, job_or_jobs, opts)
     |> Map.put(:url, "#{@base_url}/jobs/delete.json")
-    |> make_request()
+    |> Helpers.make_request()
   end
 
   def stop(api_key, project_id, job_or_jobs, opts \\ [])
       when is_api_key(api_key)
-      when is_project_id(project_id)
-      when is_job_id(job_or_jobs) or is_list(job_or_jobs)
+      when is_id(project_id)
+      when is_id(job_or_jobs) or is_list(job_or_jobs)
       when is_list(opts) do
     prepare_basic_request(api_key, project_id, job_or_jobs, opts)
     |> Map.put(:url, "#{@base_url}/jobs/stop.json")
-    |> make_request()
+    |> Helpers.make_request()
   end
 
   defp prepare_basic_request(api_key, project_id, job_or_jobs, opts) do
@@ -111,16 +110,6 @@ defmodule SHEx.Endpoints.App.Jobs do
     job_or_jobs
     |> List.wrap()
     |> Enum.map(&{:job, &1})
-  end
-
-  defp make_request(%RequestConfig{opts: opts} = config) do
-    http_client = get_http_client(opts)
-
-    config |> http_client.request()
-  end
-
-  defp get_http_client(opts) do
-    opts |> Keyword.get(:http_adapter, DefaultAdapter)
   end
 
   defp format_job_settings(nil, _encoder), do: {:ok, []}
@@ -149,18 +138,6 @@ defmodule SHEx.Endpoints.App.Jobs do
 
   defp maybe_add_job_settings(list, settings) do
     list |> Keyword.put(:job_settings, settings)
-  end
-
-  defp validate_params(params, expected) do
-    params
-    |> Enum.reject(fn {k, _} -> Enum.member?(expected, k) end)
-    |> case do
-      [] ->
-        :ok
-
-      invalid_params ->
-        {:invalid_params, {Keyword.keys(invalid_params), "valid params: #{inspect(expected)}"}}
-    end
   end
 
   defp validate_state(nil), do: :ok
